LMS-8 refactor: change file
